backend/pyproject.toml
[tool.poetry]
name = "emby-next-api"
version = "0.1.0"
description = "Emby Next Gen Console - API"
python = "^3.12"

[tool.poetry.dependencies]
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
pydantic = "^2.7"
pydantic-settings = "^2.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.29"
alembic = "^1.13"
redis = {extras = ["hiredis"], version = "^5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7"}
pyjwt = "^2.8"
gunicorn = "^22.0"